+++ zk_toolbox/crates/zk_supervisor/src/commands/prover/insert_batch.rs
